+++ src/test/kotlin/racingcar/model/CarNameParserTest.kt
@@ -4,7 +4,7 @@ import camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest
 import camp.nextstep.edu.missionutils.test.NsTest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import racingcar.controller.main
+import racingcar.application.main
 
 class CarNameParserTest : NsTest() {
